resnet: switch to own resnet impl
It was too annoying to try to subclass the BasicBlock
and figure out why the model was unhappy with the code. Also,
this lets us more easily fix up the block to do preactivation
into conv as discussed in the paper instead of the default
impl in pytorch.

This one now lets us do arbitrary k args.